requirements.txt
asgiref==3.7.2
channels==4.0.0
channels-redis==4.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
msgpack==1.0.7
packaging==23.2
Pillow==10.1.0
pip-review==1.3.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
